added UserProfile model
